+++ src/SubwidgetContainer.cpp
@@ -226,10 +226,14 @@ namespace tgui
 
         Widget::rendererChanged(property);
 
-        // Make sure to pass on the SubwidgetContainer's font to the internal container
+        // Make sure to pass on the SubwidgetContainer's font, opacity and text size to the internal container
         if (property == U"Font")
         {
             m_container->setInheritedFont(m_fontCached);
+        }
+        else if ((property == U"Opacity") || (property == U"OpacityDisabled"))
+        {
+            m_container->setInheritedOpacity(m_opacityCached);
         }
     }